+++ index.html
@@ -12,7 +12,7 @@ html, body {
   background-size: 20px 20px;
 }
 
-.hover-help {
+.hover-debug-helper {
   font-family: "Consolas", "Monaco", monospace;
   color: white;
   background: rgba(0, 0, 0, 0.75);
